requirements.txt
streamlit==1.32.0

langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-huggingface==0.0.3
langchain-groq==0.1.5

faiss-cpu==1.7.4
sentence-transformers==2.7.0

python-dotenv==1.0.1

flashrank==0.2.5